# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Product Service",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Log in with username and password",
                "tags": [
                    "User"
                ],
                "summary": "Log in",
                "parameters": [
                    {
                        "description": "User",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthorizationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "description": "Sign up a new user with username and password",
                "tags": [
                    "User"
                ],
                "summary": "Signing up new user",
                "parameters": [
                    {
                        "description": "User",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user_service.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthorizationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "description": "Get a list of all categories",
                "tags": [
                    "Category"
                ],
                "summary": "Get a list of all categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
        "/api/categories/create": {
            "post": {
                "description": "CreateNewProduct with json struct",
                "tags": [
                    "Category"
                ],
                "summary": "Creating New Category",
                "parameters": [
                    {
                        "description": "Category",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/category_service.CategoryCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
        "/api/categories/update": {
            "patch": {
                "description": "Update any field / fields of category\nOnly ID is required",
                "tags": [
                    "Category"
                ],
                "summary": "Updating category information",
                "parameters": [
                    {
                        "description": "Category",
                        "name": "product",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/category_service.CategoryUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categories.CategoryUpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "auth.AuthorizationResponse": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "auth.UserLoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "categories.CategoryCreateResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "categories.CategoryList": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/category_service.UserCategory"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "categories.CategoryUpdateResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "category_service.CategoryCreateDTO": {
            "type": "object",
            "properties": {
                "api_name": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
        "category_service.CategoryUpdateDTO": {
            "type": "object",
            "properties": {
                "api_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
        "category_service.UserCategory": {
            "type": "object",
            "properties": {
                "api_name": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                }
            }
        },
...
        "user_service.UserDTO": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}